+++ testapps/jpa_model_id/src/main/java/jpa_test/AppEntry.java
@@ -1,4 +1,4 @@
-package test;
+package jpa_test;
 
 import act.Act;