+++ testapps/jpa_model_id/src/main/java/jpa_test/endpoint/PublicServiceBaseV1.java
@@ -1,4 +1,4 @@
-package test.endpoint;
+package jpa_test.endpoint;
 
 import org.osgl.aaa.NoAuthentication;