+++ testapps/jpa_model_id/src/main/java/jpa_test/model/Audit.java
@@ -1,4 +1,4 @@
-package test.model;
+package jpa_test.model;
 
 import act.aaa.model.AuditBase;
 import org.osgl.aaa.Principal;